requirements.txt
pandas
numpy
scipy
implicit[gpu]
scikit-learn
pyyaml
tqdm
psutil
wandb
cupy-cuda12x
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
mlflow
torch
torchvision
torchaudio